requirements.txt
openai>=1.30.0
sendgrid>=6.11.0
requests>=2.31.0
python-dotenv>=1.0.0
